+++ src/index.ts
@@ -20,6 +20,7 @@ export { MySqlGrammar } from "./schema/grammars/MySqlGrammar.js";
 export { PostgresGrammar } from "./schema/grammars/PostgresGrammar.js";
 
 export { Builder, CursorPaginator, Paginator, SimplePaginator } from "./query/Builder.js";
+export type { LikeOptions } from "./query/Builder.js";
 export type { CursorPaginatorJson, NumericAggregate, PaginatorJson, SimplePaginatorJson } from "./query/Builder.js";
 export { DB } from "./query/DB.js";
 export { Collection, collect } from "./support/Collection.js";